Ajout de delete multiples

+++ app/Http/Controllers/EvenementController.php
@@ -3,12 +3,21 @@
 namespace App\Http\Controllers;
 
 use App\Evenement;
+use Illuminate\Support\Str;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\Validator;
 
 class EvenementController extends Controller
 {
+    /**
+     * Tous les évenements
+     *
+     * @return void
+     */
+    public function index()
+    {
+        return Evenement::all();
+    }
+
     /**
      * Store a newly created resource in storage.
      *
@@ -17,17 +26,18 @@ class EvenementController extends Controller
      */
     public function store(Request $request)
     {
-        Validator::make($request->all(), [
-            'title' => 'required|string|min:3|max:50',
-            'nom' => 'required|string|min:3|max:30',
-            'lieu' => 'required|string|min:3|max:30'
-        ])->validate();
+        $request = Evenement::validPosts($request);
+        if($request == false)
+            return response()->json('Erreur de requète !', 404);
 
-        $hash = uniqid();
+        // uniqid('', true) Renvoie un hash de ce type : 5f6da1345f3230.40143896
+        $hash = Str::replaceFirst('.', '', uniqid('', true));
         $request['hash'] = $hash;
-        $hash = uniqid();
+
+        $hash = Str::replaceFirst('.', '', uniqid('', true));
         $request['hash_admin'] = $hash;
-        $evenement = Evenement::create($request->all());
+
+        $evenement = Evenement::create($request);
 
         return response()->json($evenement, 201);
     }
@@ -40,8 +50,46 @@ class EvenementController extends Controller
     public function show(String $hash)
     {
         $evenement = Evenement::firstWhere('hash', $hash);
+        if($evenement == false)
+            return response()->json('Evenement introuvable !', 404);
+        $evenement->votes;
+        $i = 0;
+        foreach($evenement->dates as $date){
+            $evenement->dates[$i] = $date['date'];
+            $i++;
+        }
+        $i = 0;
+        foreach($evenement->votes as $vote){
+            $evenement->votes[$i] = $vote['nom'] . '=>' . $vote['vote'];
+            $i++;
+        }
+        return response()->json($evenement);
+    }
+
+    /**
+     * Affichage de l'évènement en mode Admin
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function showAdmin(String $hashAdmin)
+    {
+        $evenement = Evenement::firstWhere('hash_admin', $hashAdmin);
+        if($evenement == false)
+            return response()->json('Evenement introuvable !', 404);
         $evenement->votes;
         $evenement->dates;
+
+        $evenement->makeVisible(['hash_admin']); // Rend visible le champ hash_admin
+        $i = 0;
+        foreach($evenement->dates as $date){
+            $evenement->dates[$i] = $date['date'];
+            $i++;
+        }
+        $i = 0;
+        foreach($evenement->votes as $vote){
+            $evenement->votes[$i] = $vote['nom'] . '=>' . $vote['vote'];
+            $i++;
+        }
         return response()->json($evenement);
     }
 
@@ -51,10 +99,17 @@ class EvenementController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, String $hash)
+    public function update(Request $request, String $hashAdmin)
     {
-        $evenement = Evenement::firstWhere('hash', $hash);
-        $evenement->update($request->all());
+        $request = Evenement::validPosts($request);
+        if($request == false)
+            return response()->json('Erreur de requète !', 404);
+
+        $evenement = Evenement::firstWhere('hash_admin', $hashAdmin);
+        if($evenement == false)
+            return response()->json('Evenement introuvable !', 404);
+
+        $evenement->update($request);
         return response()->json($evenement);
     }
 
@@ -63,10 +118,24 @@ class EvenementController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function destroy(String $hash)
+    public function destroy(String $hashAdmin)
     {
-        $evenement = Evenement::firstWhere('hash', $hash);
-        $evenement->delete();
-        return response()->json('OK', 204);
+        $evenement = Evenement::firstWhere('hash_admin', $hashAdmin);
+        if($evenement == false)
+            return response()->json('Evenement introuvable !', 404);
+
+        $evenement->delete(); // Avec les cascades, les créneaux et les votes seront supprimés en même temps
+        return response()->json('Suppréssion OK', 204);
+    }
+
+    public function destroyAll()
+    {
+        $evenements = Evenement::all();
+        foreach ($evenements as $evenement) {
+            $evenement = Evenement::findOrFail($evenement->id);
+            $evenement->delete();
+        }
+
+        return response()->json('Suppréssions multiples OK', 204);
     }
 }
